switched aggr->values to uint32_t to keep consistency. This needs to be uint64_t though.

+++ src/engine/include/pipeline.h
@@ -150,7 +150,7 @@ struct groupfilter_result {
 
 struct aggr {
   size_t                                      num_values;
-  uint64_t*                                   values;
+  uint32_t*                                   values;
 };
 
 /* -----------------------------------------------------------------------*/
